divide game_basic_structs.h to card.h and deck.h
