+++ card.cpp
@@ -1,0 +1,29 @@
+#include "card.h"
+
+Card::Card() {}
+
+Card::Card(int _suit, int _rank) {
+	if (suits.find(_suit) != suits.end()) {
+		this->suit = suits.at(_suit);
+	}
+	else {
+		std::cout << "This suit doesn't exist!";
+	}
+
+	if (ranks.find(_rank) != ranks.end()) {
+		this->rank = ranks.at(_rank);
+	}
+	else {
+		std::cout << "This rank doesn't exist!";
+	}
+}
+
+std::string Card::getSuit() { return this->suit; }
+
+std::string Card::getRank() { return this->rank; }
+
+std::ostream& operator<<(std::ostream& out, Card& card) {
+	return out << card.getRank() << " " << card.getSuit() << std::endl;
+}
+
+Card::~Card() {}
