+++ card.h
@@ -39,23 +39,3 @@ public:
 
 	~Card();
 };
-
-
-
-
-
-class Deck
-{
-private:
-	
-public:
-	static Deck* Instance;
-
-	Deck();
-
-	std::vector<Card> deck;
-
-	void shuffling();
-
-	~Deck();
-};
